Move json directory out of templates
