+++ packages/cli/src/init/json/biome.json
