+++ packages/cli/src/init/json/kv.json
