+++ packages/cli/src/init/json/mq.json
