+++ packages/cli/src/init/json/pm.json
