+++ packages/cli/src/init/json/rt.json
