+++ packages/cli/src/init/json/vscode-settings-for-deno.json
