+++ packages/cli/src/init/json/vscode-settings.json
